Move creatable scheduled task types to /models
