+++ src/models/scheduled-task/index.ts
@@ -6,13 +6,21 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
+export * from './creatable-scheduled-query';
+export * from './creatable-scheduled-scripts';
+export * from './creatable-scheduled-task';
+export * from './creatable-scheduled-task-base';
 export * from './is-scheduled-query';
 export * from './is-scheduled-script';
 export * from './is-scheduled-task-base';
+export * from './raw-creatable-scheduled-task';
 export * from './raw-scheduled-task';
 export * from './scheduled-script';
 export * from './scheduled-task';
 export * from './scheduled-task-base';
 export * from './squeduled-query';
+export * from './tagged-creatable-scheduled-query';
+export * from './tagged-creatable-scheduled-script';
+export * from './to-raw-creatable-scheduled-task';
 export * from './to-scheduled-task';
 export * from './to-scheduled-task-base';
